initial move to own lib
